cincinnati/lib: add support for deserialization
This will allow client code to consume graphs sent over the network.
It's worth noting that Release had to be reordered to get Serde to try
to deserialize as a ConcreteRelease before falling back to an
AbstractRelease. This is only necessary because the representation of
Release is untagged and the fields of AbstractRelease are as strict
subset of ConcreteRelease.