graph-builder: improve logging output
Instead of logging the layer contents (which can be very large), it's
more useful to log the digest of the layer. Additionally, if no layers
contain the metadata, it's more clear to express that in the error
message instead of returning the last error (which indicates that a
particular layer didn't have the metadata).