*: bugfix and refactor prefix semantics
The bugfix and semantic change is that the prefix keeps, and if
necessary adds, a leading slash. This simplifies the logic of dealing
with prefixes because it doesn't require special-casing for emtpy
namespaces which was buggy before.

Using actix-web's prefix method removes the need to specify the prefix
for every new route.