Revert "*: rename payload -> image"
This reverts commit 5d226dab3ca1a45cea13b2b6753cb08f6f05cb0d.

After some discussion, we've decided that the term "payload" is more
appropriate here given that not all consumers of Cincinnati will specify
their updates as an image reference. The idea of expressing the payload
as an extendable structure was also discussed. We determined that this
structure would need to remain opaque to the Cincinnati stack (since
different consumers have different requirements) and would ultimately
complicate things too much at this stage in the project.