graph-builder: use plugins for most operations
Replace the inline graph-builder code with functionality from plugins
for fetching and processing the quay metadata.

Remove the previously introduced metadata trait as the new plugin
system deprecates the need to standardize around the metadata
specifically. Common functionality in case of multiple metadata plugins
might be extracted on occasion. The metadata constants are now owned by
the quay plugin as they are quay specific.