Add prometheus port to the containers and services
Add the port that the metrics will be exposed on.
The ports are named:
- prometheus-policy-engine (port 9081)
- prometheus-graph-builder (port 9080)

Since it is required that all ports must be named in a multi-port
service, add a name to the policy-engine and graph-builder ports.

Also expose the ports via the service, so that the endpoints are
added, and servicemonitors can select them