cincinnati/plugin/internal: add channel_filter
The ChannelFilterPlugin can be used to filter a graph by a specific
channel. It reads the requested channel from the parameters value at key
"channel". The channel name is validated according to a defined regex.

If a release belongs to several channels the other channel names are
removed from the result.