openshift: use tcpSocket for policy-engine liveness-probe
When the policy-engine requires parameters on the graph endpoint the
pre-existing probes will fail as they are.  This switches to using TCP
socket probes instead which are agnostic of the required parameters.
The readiness-probe can be removed as there's no distinct condition for
readiness within the policy-engine at this point.