graph-builder: use parking_lot::RwLock and refactor
The motivation for this change is the fairness aspect of this RwLock,
which eliminates the risk for lock starvation.

The semantics of the new RwLock are significantly as there is no lock
poisoning. Adapting the code has shed some light on the status methods,
which took the lock poisoning state into consideration.
As a simplification there's only one state structure for status and
graph, owned by the latter.