docs/design/cincinnati: Remove edges instead of removing nodes
We need to keep the degraded release node in the graph so existing
clusters can upgrade away.  Removing incoming edges is sufficient to
keep new clusters off the degraded release.

Also add 1.1.1 to the Alpha graph, because we're not going to put the
patch release straight into Beta without testing in Alpha first.

For the direct 1.0.0 -> 1.2.0 edge, updating edges via Quay labels is
possible today [1].  That means that there's really not much point to
having the 'next' edge metadata, since you could always update the
target node's metadata to include a 'previous' reference.  But I've
left that alone for now.

[1]: https://github.com/openshift/cincinnati/pull/103#discussion_r284625343