cincinnati/plugins: add assign_wariness plugin
This implements a new "assign_wariness" plugin, which is mainly meant
to be used on policy-engine side.

This plugin ensures that each request has an attached "wariness" score
for throttling purposes, which can be either provided by the client
or computed server-side based on some other client input paramenters.

Wariness value can be then used by other plugins to enact (static or dynamic)
throttling decisions. It has the property that a set of request parameters
will always result in the same wariness value, for stable/deterministic
bucketing.