plugins/edge_add_remove: gracefully handle duplicate edges
The plugin shall ignore the error when attempting to add a duplicat
edge in order to not propagate the error and make the plugin chain fail
when it previously didn't.

Due to the underlying library change the result will be different and no
longer contain duplicate edges.

The introduction of this error caused some of the tests to be broken.
These were relying on the duplicate edge behavior.