cincinnati,policy-engine: extract graph-fetch code to an internal plugin
This extracts the code for fetching a Cincinnati Graph into an
InternalPlugin. This plugin is special so that it is expected to be the
first plugin in the chain. This is because it discards its input graph
and yields the graph which it fetches remotely.

The plugin needs to register its metrics with the registry of the outer
scope. Therefore the interfaces have been adjusted to plug a registry
through to the plugin instantiation.

As the error handling code of the graph-fetching could yield a
GraphError, the generic plugin processing code will now attempt to
downcast the generic failure::Error to the GraphErro type which is
expected on the graph endpoint.

More changes:
  * add a String for details to the FAiledUpstreamRequest error type
  * deprecate the upstream setting on the the policy-engine as this can
  be configured via the plugin configuration instead.