policy-engine/graph: split query params into at most 2 parts
Before this change any characters after a '=' in the value
are not passed on as a plugin parameter value. This leads to incorrect
behavior, e.g. ignoring invalid values or losing information in case the
'=' is actually valid.

The `actix_web::Query::from_query` implementaiton takes care of this
nicely and we get rid of the manual -- and undesired -- parsing code.