openshift: enforce CPU limits on each container
This introduces CPU limit settings on both graph-builder and
policy-engine, to avoid disrupting all cluster resources in case
of issues.
The default value (750 millicores) has been picked as 2x the maximum
value observed so far in production. It can be tweaked later on based
on further observations.
It also adds baseline resources request for pod scheduling.

Closes: https://jira.coreos.com/browse/APPSRE-308