cincinnati/src/plugins/internal/channel_filter: Expose all channels
This filter landed in cfa652e251 (cincinnati/plugin/internal: add
channel_filter, 2019-04-05, #79), with filtering the metadata as an
explicit goal:

  If a release belongs to several channels the other channel names are
  removed from the result.

But the channel-switching procedure is:

1. Upgrade to a release that belongs to both your current channel
   (e.g. prerelease-4.1) and the target channel (e.g. stable-4.1).
2. Change your channel.
3. Upgrade in the new channel.

This commit leaves the channel metadata alone, so now the channel
filter is only removing releases that don't belong to the channel and
is no longer adjusting metadata for releases that do belong to the
channel.  The open list won't work with secret channels, but we don't
have any short-/mid-term plans for secret channels, and we have very
short-term plans for 4.1->4.2 edges.  Tooling like the web console's
channel-picker can use the new approach to retrieve available channels
for the current release, instead of hard-coding a list of channel
choices that really has nothing to do with the current release
(e.g. [1]).

[1]: https://github.com/openshift/console/pull/1498