cincinnati/plugins/edge_add_remove: support arch
This code makes the plugin take an architecture metadata into account
when identifying source or destination release versions for edges to be processed.
The assumption is that if architecture metadata exists, it exists for
both ends of the edge which is processed, thus this plugin will find
matching architecture pairs, even if multiple versions with different
architectures are contained in the graph.