policy-engine/src/main: Add request logging
The default log format is [1]:

  %a %t "%r" %s %b "%{Referer}i" "%{User-Agent}i" %T

This will give us a user-agent to help debug broken clients.  Before
this commit, logs looked like:

  [2019-12-06T20:14:16Z ERROR policy_engine::graph] Error serving request with parameters '"channel=stable-4.1&version=4.2.4"': mandatory client parameters missing
  [2019-12-06T20:14:29Z ERROR policy_engine::graph] Error serving request with parameters '"channel=stable-4.2&version=4.2.4"': mandatory client parameters missing
  [2019-12-06T20:14:52Z ERROR policy_engine::graph] Error serving request with parameters '"channel=fast-4.2&version=4.2.4"': mandatory client parameters missing

which makes it hard to track down the broken clients.  Ideally we'd
only include the user-agent logs for invalid requests, but I'm not
clear on how to set that up.

[1]: https://actix.rs/docs/middleware/#usage