tests/e2e: sort expected/received releases by version
Cincinnati may return unsorted list of nodes if tags are processed
in parallel. Expected nodes are sorted as well, so that expected nodes
could be autogenerated and those would not have to be sorted by version.

Also shuffle nodes in the testdata to ensure expected is sorted properly
as well.