dist/openshift/graph-data-sync: Syncer Deployment
I tried to use CentOS 8 from docker.io, since that seems to be their
canonical container image [1].  We only need Python, Bash, tar, and
POSIX utilities.  But jobs launched with that base failed with:

  env: 'python': No such file or directory

so now I'm using Docker's Python container.

Setting HOME avoids:

  Installing collected packages: PyYAML
  ERROR: Could not install packages due to an EnvironmentError: [Errno 13] Permission denied: '/.local'

Docs on Templates in [2], but if you don't want to read through all of
that, you can apply this to a cluster with:

  $ oc process -f graph-data-sync.yaml --local -o yaml | oc -n "${YOUR_NAMESPACE}" create -f -

[1]: https://www.centos.org/download/
[2]: https://access.redhat.com/documentation/en-us/openshift_container_platform/4.3/html/images/using-templates