cincinnati,graph-builder: move plugins, extend catalog, make plugins configurable
The primary goal of this change is to make the graph-builder plugins
configurable via a config file.

This required all of the configurable plugins to be in the plugin
catalog, which lives in the cincinnati crate.  Therefore all
graph-builder plugins are moved to the cincinnati crate, leaving the
graph-builder crate with only its core functionality.

Subsequently the graph-builder is made configurable, with a few
refactors along the way as they fit in naturally with this change:
* and the naming of the related functions and variable names in the
  graph-builder and policy-engine are normalized.
* the yaml_lint