commons/testing: add sort_json_graph_by_version
This function is copied from the e2e testsuite to be reused by other
components. It's not a verbatim copy, as it adds a fix to also sort the
edges array in itself. This is required to make the output fully
predictable in case a node has multiple edges, and thus makes the output
This assertable for equality.