cincinnati/plugins: expose and use DEFAULT_KEY_FILTER in OpenshiftSecondaryMetadataParser
We want to use an OpenShift specific default value for the key_prefix
setting. However, using a default value for the key_prefix string which
is not the default for the string type, which would be the empty string,
is not intuitive.  This can be alleviated by making the default value a
publicly expose constant and thus make it transparent.