dist/openshift/cincinnati: Drop graph-builder pause to 5 minutes
The default seems to have lived in the 30s range since at least
517af6ee8f (graph-builder: add support for configuration file,
2019-03-12, #69).  But we can allow some latency here.  On the
cluster-version operator (CVO) side:

* cvo.New takes an argument for the poll interval [1].
* That is fed by resyncPeriod(o.ResyncInterval)() [2].
* ResyncInterval is set from minResyncPeriod [3].
* minResyncPeriod is 2 minutes [4].
* resyncPeriod adds some jitter (returning between 1x and 2x the input
  value) [5].

So should be uniformly distributed between 2 and 4 minutes.  That puts
a floor on latency in getting this information out to clusters.

It's also not a problem to add latency above the CVO-enforced floor.
A delay in minutes in getting a new release out into a channel is not
significant, because the biggest publication rush would be releasing a
CVE fix into the stable channel.  And that's already delayed by
needing to cook in the candidate and fast channels, and eventually
having the push to stable being a phased rollout.

A delay in pulling an edge that has proven itself unstable is more
serious, but we already have latencies in the minutes before alerts
fire, and Telemetry/Insights that report issues to Red Hat add
additional latencies in the minutes (Telemetry) to hours (Insights)
range.  Delaying the edge pull by a few more minutes is not going to
greatly increase overall cluster-breaks -> edge-pulled latency here.

The above discussion hopefully shows that there is little downside to
growing the latency up to several minutes.  With this commit, the
delay is 5 minutes plus the time it takes the plugins to execute (I'd
initially argued for 10m, but was bargained back down to 5 [6]).  The
upside is that it decreases our load on our external dependencies
(quay.io, GitHub), reducing issues like:

  [2020-03-31T21:35:16Z DEBUG graph_builder::graph] graph update triggered
  [2020-03-31T21:35:20Z ERROR graph_builder::graph] Checking for new commit
  [2020-03-31T21:35:20Z ERROR graph_builder::graph] Parsing {"message":"API rate limit exceeded for user ID ...","documentation_url":"https://developer.github.com/v3/#rate-limiting"} to Vec<Branch>
  [2020-03-31T21:35:20Z ERROR graph_builder::graph] invalid type: map, expected a sequence at line 1 column 0
  [2020-03-31T21:35:50Z DEBUG graph_builder::graph] graph update triggered

It's also reduces our CPU and network load, since the graph-builder
pod spends more time resting in between recalculate-the-graph spikes.
Just the load consideration is enough for me to want to tune this
latency as high as we can while keeping it well below other
edge-pulling latency components.

[1]: https://github.com/openshift/cluster-version-operator/blob/12623332615a0657a5468fae27cff1998a70bfef/pkg/cvo/cvo.go#L165
[2]: https://github.com/openshift/cluster-version-operator/blob/12623332615a0657a5468fae27cff1998a70bfef/pkg/start/start.go#L341
[3]: https://github.com/openshift/cluster-version-operator/blob/12623332615a0657a5468fae27cff1998a70bfef/pkg/start/start.go#L93
[4]: https://github.com/openshift/cluster-version-operator/blob/12623332615a0657a5468fae27cff1998a70bfef/pkg/start/start.go#L43
[5]: https://github.com/openshift/cluster-version-operator/blob/12623332615a0657a5468fae27cff1998a70bfef/pkg/start/start.go#L250-L251
[6]: https://github.com/openshift/cincinnati/pull/264#discussion_r402880161