plugins/scrape_dockerv2/tests: move tests from registry to plugin
This moves all tests from the submodule to the plugin level. This allows
usage of test related functionality which exists for plugin-level tests.
The trade off to abstract registry-level details from the tests here is
welcome as we ultimately want to test for the plugin result and not the
lower level details.