[build] remove dependency on dockerfile.build
we are doing a multi-stage build in dockerfile.deploy to
remove the dependency on dockerfile.build

this is because we are installing certain modules in dockerfile.build
that are introducing vulnerabilities. We are not carrying these
vulnerabilities in the final container that we are shipping, we are just
using dockerfile.build container as a builder image. This can be achieved
in a multi-stage build, thus not requiring a separate build dockerfile.
dockerfile.build will be removed in a separate commit.