fix manifestref error if digest value missing
Currently if a ConcreteRelease is missing the manifestref digest,
Cincinnati tries to unwrap() this value which causes it to panic
becuase of the absense of digest.

This commit returns an invalid value for manifestref digest if
it is missing in the ConcreteRelease.
This will cause a mismatch in fn add_release() which will prevent
cincinnati from adding this release. Thus preventing the panic
and throwing an easily readable error.