do not panic on sha-mismatch for multi-arch images
check the architecture of images before panicking for sha
mismatch.
this is needed as the multi-arch images for single-arch
have same version as single-arch images but a different sha.
multi-arch images for single-arch contains additional information
and are different from the images built specifically for single-arch

with this commit, cincinnati will ignore single-arch tagged images
from multi-arch manifest and will only keep single-arch images in
the graph.