OTA-920: fix permission issue while extracting containers
This bumps dkregistry to the latest commit and uses
dkregistry::render::filter_unpack to only extract the files which
match the output allowlist.  As a bonus, this removes the code to
remove disallowed files since it is no longer needed; the disallowed
files are never unpacked onto the filesystem in the first place.