revert mockito to older version and fix tests
fix failing tests and reverts mockito dependency to older version.
mockito had a major rewrite and changes the way it functions,
which will require a major refactor for tests in cincinnati wherever
mockito dependency is used.
We might have to either replace the dependency with something similar
or refactor the tests.
The issue with newer version of mockito is that individual tests pass
but when all the tests are run simaltaneously, the tests might use
another instance of the mock which gives us incorrect and inconsistant
mock