OCPBUGS-25833: strict filtering of BlockedEdges
strictly filter the old blocks so that conditional blocks are handled
in the error section the way we were handling old blocks while parsing
conditional edges.

A blocked edge will now appear either in old_blocks or
conditional_blocks. Earlier all the conditional edges used to
appear in the old_blocks as well.