ignore incorrect images instead of erroring out
we're ignoring the images which we do not understand with this change.
this change is required because we dont want cincinnati to error out on encountering
cosign signatures.
unintended consequence of this change can be that cincinnati continues with a log when it encounters
incorrect image whereas it should error out. Cincinnati won't include this image in the update graph.
We should ideally try to ignore only the cosign .sig signatures. this change acts a
stopgap till we teach cincinnati to deal with signatures

log like this will be printed when we encounter a incorrect images
[2024-06-05T20:04:15Z WARN  cincinnati::plugins::internal::graph_builder::release_scrape_dockerv2::registry] error fetching manifest and manifestref for <>/<>-releases:sha256-7c42e086007d868f8295692.sig: fetching manifest and manifestref for <>/<>-releases:sha256-7c42e086007d868f8295692.sig: unknown media type ManifestV2S1, ignoring this image