Return `400 Bad Request` for responses that contain empty graph
Return `400 Bad Request` for responses that contain an empty graph. Such
responses are currently only generated when a specified channel and
architecture result in an empty update graph due to the filtering.

This commit is part of the effort to improve the UX regarding requests
that contain invalid query parameters. In the future, more fine-grained
return messages will be returned.