dist/cargo_test.sh: Fix various issues
dist/cargo_test.sh: Fail on compile errors

Currently, the `cargo_test.sh` script skips compilation errors of tests,
subsequently does not run the non-existing tests, and exits with a zero
exit code, as no tests failed technically.

On the line:
```
mapfile -t tests < <(
    cargo test --no-run --message-format=json ${cargo_test_flags[${directory}]} | \
    jq -r 'select(.profile.test == true) | .executable'
)
```

the `cargo test` compiles tests. However, when a compilation fails, this
does not cause the script to fail. The exit code is ignored, and the
output is piped to the `jq` command, which continues with a non-zero
code as it parses the json message successfully. However, as no tests
exist, the script continues.

This commit will compile the tests and check for the exit-code of a
compilation. The `cargo test --no-run` command is used to provide
human-readable output of compilation errors. On a non-zero exit code,
the overall script fails immediately. Originally, the script failed on a
first failed test, this commit will ensure the script fails on a first
failed test or a first compilation error.

dist/cargo_test.sh: Fail on any failed test

Fix the script to fail on any failed test. Currently, the script fails
only when the last element in the `tests` array exists with a non-zero
exit code.

Refactor the compilation of tests to fail the script using
the introduced `-e` option.

dist/cargo_test.sh: Run all tests before exiting due to failure

dist/cargo_test.sh: Simplify `run_tests`