dist/cargo_test.sh: Remove coverage, create junits
Migrate the script to use the Nextest runner (https://nexte.st) which is
able to produce JUnit XMLs right away. It also has nicer output than
standard one.

When run locally, the script falls back to standard runner when Nextest
is not present. In CI, the script fails without Nextest (we do not want
CI to silently stop producing JUnits).