increment `http_upstream_reqs` on errors
Previously the `http_upstream_reqs` was only incremented when `cached_graph`
call was successful (and result was not cached). When `cached_graph` actually
made the call internally but it was not successful, the `.await?` shortcut
returned the error right away and the counter was never incremented, even
when the call was actually made.

This is actually tested in the `fetch_upstream_failure_test` macro:

```rust
assert_eq!(1, plugin.http_upstream_reqs.get() as usize);
```

These tests are failing for some time and it is hard to imagine how they
were ever passing.