cincinnati_graph_fetch: simplify tests
The `clone()` calls are `Arc::clone()` tests under the hood, so they
result in returning smart pointers. This is legit (we still `get()` on a
correct thing when testing) but it does not seem to be necessary and it
is confusing (how are we expecting a clone to be incremented? aha, they
are not actually clones under the hood!) so we can remove them