deps: remove unused `failure` dependency
It seems that this crate was used in the past to provide `Error` traits
implementation, but was eventually replaced with `thiserror::Error`
(see the removed `use thiserror::Error as Fail` rename). The `Fail`
trait was only derived at several places, so I dropped the rename
and used `thiserror::Error` directly in derives. The `Fail` from the
commons package was also used in the cincinnati package, relying on a
transitive dependency when the code is actually directly used. I dropped
this indirection and made cincinnati package depend on thiserror (same
version like commons)