Fix metadata parser test fixture
The commit [1] has removed conditional edges from normal edges. This was
done using the `io.openshift.upgrades.graph.previous.remove_regex`
metadata key. The `from` regex field of a conditional risk is appended
to the key's value of a respective `to` version. In other words, the
`OpenshiftSecondaryMetadataParser` parser signals using the metadata
key to the `EdgeAddRemove` plugin what edges to remove from the normal
edges thanks to its knowledge of conditional risks.

Unfortunately, a test fixture was not modified to adhere to the new
changes. The `graph-gb-with-quay-metadata.json` file is supposed to
represent graph with the metadata to act as control data.
This commit fixes that.

[1] https://github.com/openshift/cincinnati/commit/72ed35630cfcb0796473deffdc6c6c2c31ed1155