graph_builder: Prepend http scheme to registry URLs that miss scheme while parsing
Before this change, the `example` scheme was prepended to a registry URL
that did not specify a scheme. Unfortunately, this has resulted in the
URL parsing library to creating different `path` values for different
similar URLs without a namespace. The `path` value is propagated to a
registry `namespace` property. And such, some URLs would have
a namespace `""` and some would have the namespace `"/"`. Now all such
URLs will result in the namespace `"/"`.

The current library processes the `http` scheme as *non-special* and
as such, it does a different processing. For the cincinnati, it means that
the library will now process all registry URLs missing a scheme the same
way. This will simplify testing.