graph_builder: Fix and extend registry URL parser tests
A test for host_port_string method was present. However, it was
checking its result against a value that even contained the
namespace. This was resulting in the tests to fail.

The commit introduces tests for host_port_string and
host_port_namespaced_string methods and sets them up correctly.