dist/openshift: Set maxUnavailable:0 in Deployments
We've had this since the Deployment template was created in 55ea56e9b4
([dist] add deployment as an alternative to deploymentconfig,
2023-03-02, #785).  That was probably just pinning Kubernetes' default
value [1].

But Cincinnati pods come up slowly, and we seem to have trouble coming
back out of overwhelmed-by-traffic holes.  Dropping maxUnavailable to
zero will force us to surge into new configuration, ensuring we always
have new-config capacity available before we reduce outgoing-config
capacity.

The risk with this approach is if there's no room for new-config
capacity, but would be room for a more in-place roll.  But we don't
have to worry about that most of the time in our app-interface because
we're hosted on autoscaling compute:

  app-interface $ grep -A19 machinePools: data/openshift/appsrep08ue2/cluster.yml | sed 's/\(subnet-\).*/\1REDACTED/'
  machinePools:
  - id: workers-0
    instance_type: m5.2xlarge
    subnet: subnet-REDACTED
    autoscale:
      min_replicas: 1
      max_replicas: 4
  - id: workers-1
    instance_type: m5.2xlarge
    subnet: subnet-REDACTED
    autoscale:
      min_replicas: 1
      max_replicas: 4
  - id: workers-2
    instance_type: m5.2xlarge
    subnet: subnet-REDACTED
    autoscale:
      min_replicas: 1
      max_replicas: 4

There is still a risk that the cloud provider is temporarily out of
our target instance type or whatever, but:

* Machine/Node pools failing to scale could already be an issue, and
  alerting on that is orthogonal to this change.
* Deployment rollouts failing to progress could already be an issue,
  and alerting on that is orthogonal to this change.

[1]: https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#max-unavailable