dkrv2 scraper: Use ubi9-minimal base image in test Dockerfile
The previously used quay.io/app-sre/busybox base container image is not
easily accessible and requires authorization. Only the `tar` command is
needed from the base image.

Instead, use an accessible base container image ubi9-minimal;
authorization is not needed, and is from a reputable source.

As the image does not contain the necessary tolling, download it.

The base image is quite large for its purpose; however, the image is
only used in the build stage and its content is not part of the final
image.