dkrv2_openshift_secondary_metadata_scraper: Use accessible base image in test Dockerfile
The previously used quay.io/app-sre/busybox base container image is not
easily accessible and requires authorization. Only the `tar` command is
needed from the base image.

Instead, use an accessible base container image
registry.access.redhat.com/ubi9/ubi, which contains the necessary
tooling, authorization is not needed, and is from a reputable source.

The base image is quite large for its purpose; however, the image is
only used as a builder image and its content are not part of the final
image.