dkrv2 scraper: Fix default allowlist to allow required files
Otherwise, nothing is exported from a graph data container image as
nothing matches the regexes created based on the default allowlist.

The regex are used as predicates in the `filter_unpack` function [1]
to filter files. However, the predicates are used against relative
paths of the archive entries from the archive perspective.

For example, `raw/metadata.json` and not `/raw/metadata.json`.

Thus, it is needed to remove the leading slash in the default allowlist
regexes as their existence will cause the plugin to not match anything.

[1] https://github.com/camallo/dkregistry-rs/blob/3e242ee9e39646da6ff4a886e080085cc1810d37/src/render.rs#L27-L52