dkrv2 scraper: Fix default allowlist to not allow unnecessary files
Otherwise, the following unnecessary files are extracted as well:
- the files inside the internal-channels directory [1]
- some of the system files [2]

Modify tests to compare the regexes against extracted paths stripped of the
prefix of the directory they are included in. As the regexes now contain the
`^` symbol to match from the beginning, it is needed to remove the data directory
prefix from the extracted paths strings. Otherwise, the test fails even though
the correct files were extracted.

[1] https://github.com/openshift/cincinnati/commit/97dc7615f185d74c2c670ff8681a4ba18973322e
[2] https://github.com/openshift/cincinnati/pull/800/commits/9d6e9d9146b16103a997003a9aa018e8dd8d455d