dkrv2_openshift_secondary_metadata_scraper: Add tests for default allowlist regexes
To ensure required files are always matched, and the unnecessary files are not matched.

Otherwise, a regex regression is able to easily slip by into the code.
For example, the ``"channels/.+\\.ya+ml$"` regex pattern matches even files
in the `internal-channels` directory. No other existing tests are able to discover
a similar regression.

An alternative is to compare extracted files against expected files. This would require
more effort, and thus this commit aims to test the default regexes patterns only.