dist/openshift/*-deployment: Fix rollingParams to rollingUpdate
Avoiding:

  warning: error calculating patch from openapi spec: unable to find api field "rollingParams" in io.k8s.api.apps.v1.DeploymentStrategy

because the Deployment config property is rollingUpdate [1].
rollingParams was the DeploymentConfig property [2], and this
transition just got overlooked during the 55ea56e9b4 ([dist] add
deployment as an alternative to deploymentconfig, 2023-03-02, #785).
Happily, changes to the way AppSRE deploys patches or to the Kube API
server handling in our host cluster is now warning us about this kind
of unrecognized property name, and we can fix the typos in commits
like this.

[1]: https://kubernetes.io/docs/concepts/workloads/controllers/deployment/#rolling-update-deployment
[2]: https://docs.openshift.com/container-platform/4.17/applications/deployments/deployment-strategies.html#deployments-rolling-strategy_deployment-strategies