dist/cargo_test.sh: Do not fail when testing rh-manifest-generator using nextest
The rh-manifest-generator package does not have any tests.
By default, nextest fails if there are no tests to run.

As it is expected for the rh-manifest-generator package to have no tests
at the moment, pass when running `cargo nextest run` on the package.
Fail on other packages, as it is expected for them to contain tests.

`cargo test` passes on empty tests. To be more consistent across the
test runners, we should pass if there are no tests to run. However, as
we utilize the nextest in CI, I would like to be more strict when using
the runner if easily possible.